+++ compile.sh
@@ -1,0 +1,2 @@
+python3.14 -m pip install nuitka
+python3.14 -m nuitka --onefile --include-package=textual --include-package=rich main.py